requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0

# Data collection
yfinance>=0.2.0
requests>=2.31.0

# Technical analysis
ta>=0.10.0
talib-binary>=0.4.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0

# Monte Carlo & Statistics
scipy>=1.11.0

# Configuration
pyyaml>=6.0.0

# Async processing
aiohttp>=3.8.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
